Aug19 can edit staff account, can open page for employess but cannot edit

+++ db/migrate/20250812005749_remove_hashed_password_from_admin_members.rb
@@ -1,5 +1,5 @@
 class RemoveHashedPasswordFromAdminMembers < ActiveRecord::Migration[6.1]
   def change
-    remove_column :admin_members, :hashed_password, :string
+    remove_column :admin_members, :hashed_password, :string if column_exists?(:admin_members, :hashed_password)
   end
 end
